\name{doublewells-package}
\alias{doublewells-package}
\alias{doublewells}
\docType{package}
\title{
\packageTitle{doublewells}
}
\description{
\packageDescription{doublewells}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{doublewells}
\packageIndices{doublewells}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{doublewells}

Maintainer: \packageMaintainer{doublewells}
}
\references{
%~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
